CRUD working

+++ types/student.go
@@ -1,7 +1,7 @@
 package types
 
 type Student struct {
-	StudentID      int    `json:"student_id"`
+	StudentID      int64  `json:"student_id"`
 	Rut            string `json:"rut"`
 	Name           string `json:"name"`
 	Representative string `json:"representative"`
